% Auto-generated: do not edit by hand
\name{''KPIWidget}

\alias{''KPIWidget}

\title{KPIWidget component}

\description{
Component description
}

\usage{
''KPIWidget(id=NULL)
}

\arguments{
\item{id}{Character. Unique ID to identify this component in Dash callbacks.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

