% Auto-generated: do not edit by hand
\name{''KPIWidget}

\alias{''KPIWidget}

\title{KPIWidget component}

\description{
Component description
}

\usage{
''KPIWidget(id=NULL, enabled=NULL, inverted=NULL, name=NULL,
preValue=NULL, target=NULL, units=NULL, value=NULL,
variance=NULL)
}

\arguments{
\item{id}{Character. Unique ID to identify this component in Dash callbacks.}

\item{enabled}{Logical. }

\item{inverted}{Logical. }

\item{name}{Character. }

\item{preValue}{Numeric. }

\item{target}{Numeric. }

\item{units}{Character. }

\item{value}{Numeric. }

\item{variance}{Logical. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

